tools/CLI-tool/requirements.txt
typer==0.9.0
pyyaml
GitPython>=3.1.0
requests
uv
